test submission

+++ homeworks/hw2/Blog/src/main/java/com/hw2/blog/service/PostService.java
@@ -21,7 +21,7 @@ public class PostService {
         return blog_list;
     }
 
-    public List<Post> getLikedPosts(long profile_id){
+    public List<Post> getLikedPosts(Integer profile_id){
         return (List<Post>) postRepository.getLikedPosts(profile_id);
     }
